Define ResetPyvista outside conf.py
